dev-requirements.txt
networkx>=2.8.4
stellargraph>=1.2.1
chardet>=5.0.0
tqdm>=4.64.0
pytest==7.1.2
codecov==2.1.12
pytest-cov==3.0.0
twine==4.0.1
sphinx
sphinx_gallery
sphinx_rtd_theme
sphinx-copybutton
numpydoc
nbsphinx